clean up proxy for nxos to be more stable
